Update surfshark.md
Added step by step details covering how to obtain the correct WireGuard connection details. This is catching people out due to the IP address / Server Address shown on screen on Surfshark's Web page are not the ones to use with GlueTUN.

Separated the OpenVPN and WireGuard parts of the page  to try to make it easier to follow.

Fixed a typo in the Multi hop regions.

Update surfshark.md

Fixed my own typo